requirement.txt

aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
arxiv==1.4.8
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
feedparser==6.0.10
Flask==2.3.2
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
psycopg2-binary==2.9.7
PyJWT==2.7.0
PyMuPDF==1.22.5
PyPDF2==3.0.1
pytz==2023.3
requests==2.31.0
sgmllib3k==1.0.0
smmap==5.0.0
soupsieve==2.4.1
tqdm==4.66.1
twilio==8.5.0
urllib3==2.0.3
watchdog==3.0.0
Werkzeug==2.3.7
yarl==1.9.2
